pyproject.toml
[tool.poetry]
name = "bami"
version = "0.1.0"
description = ""
authors = ["grimadas <bulat.nasrulin@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pyipv8 = "2.6.0"
msgpack = "^1.0.0"
lmdb = "^"
cachetools = "^4.1.0"
mashumaro = "^2.5"
numpy = "^1.20.3"
mmh3 = "^3.0.0"

[tool.poetry.scripts]
console = "bami.console:main"

[tool.poetry.dev-dependencies]
pytest = "^5.4.2"
pytest-cov = "^2.9.0"
coverage = {version = "^5.1", extras = ["toml"]}
pytest-mock = "^3.1.0"
black = "^19.10b0"
flake8 = "^3.8.2"
flake8-bandit = "^2.1.2"
flake8-black = "^0.2.0"
flake8-bugbear = "^20.1.4"
flake8-import-order = "^0.18.1"
safety = "^1.9.0"
mypy = "^0.770"
pytype = {version = "^2020.5.13", python = "3.7"}
typeguard = "^2.7.1"
flake8-annotations = "^2.1.0"
darglint = "^1.3.0"
xdoctest = "^0.12.0"
sphinx = "^3.0.4"
sphinx-autodoc-typehints = "^1.10.3"
codecov = "^2.1.3"
pytest-asyncio = "^0.14.0"
pytest-benchmark = "^3.2.3"
pytest-xdist = "^1.33.0"
aiohttp_cors = "^0.7.0"

[tool.coverage.paths]
source = ["src", "*/site-packages"]

[tool.coverage.run]
branch = true
source = ["bami"]

[tool.coverage.report]
show_missing = true

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
